requirements.txt
Django==4.2.20
djangorestframework==3.15.2
django-filter==25.1
djangorestframework-simplejwt==5.5.0
dj-database-url==2.3.0
python-dotenv==1.0.1
requests==2.32.3
python-telegram-bot==20.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.1
